+++ packages/access-definitions/src/Wrapping/Contracts/Types/FilterableTypeInterface.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace EDT\Wrapping\Contracts\Types;
 
 use EDT\Querying\Contracts\PathsBasedInterface;
+use EDT\Querying\PropertyPaths\PropertyLink;
 
 /**
  * Defines that the implementing type can be used to filter entities of this type or other types.
@@ -38,10 +39,7 @@ interface FilterableTypeInterface extends TypeInterface
      * at all. You may also want to allow more properties for filtering than you allowed for reading,
      * but be careful as this may allow guessing values of non-readable properties.
      *
-     * @return array<non-empty-string, FilterableTypeInterface<TCondition, TSorting, object>|null> The keys in the returned array are the names of the
-     *                                   properties. Each value is the target
-     *                                   {@link TypeInterface} or `null` if the
-     *                                   property is a non-relationship.
+     * @return array<non-empty-string, PropertyLink<FilterableTypeInterface<TCondition, TSorting, object>>> The keys in the returned array are the names of the properties.
      */
     public function getFilterableProperties(): array;
 }