+++ packages/access-definitions/src/Wrapping/Utilities/SchemaPathProcessor.php
@@ -13,15 +13,15 @@ use EDT\Wrapping\Contracts\Types\FilterableTypeInterface;
 use EDT\Wrapping\Contracts\Types\TransferableTypeInterface;
 use EDT\Wrapping\Contracts\Types\SortableTypeInterface;
 use EDT\Wrapping\Contracts\Types\TypeInterface;
+use EDT\Wrapping\Properties\RelationshipReadabilityInterface;
 use EDT\Wrapping\Utilities\TypeAccessors\ExternFilterableProcessorConfig;
-use EDT\Wrapping\Utilities\TypeAccessors\ExternReadableProcessorConfig;
 use EDT\Wrapping\Utilities\TypeAccessors\ExternSortableProcessorConfig;
-use EDT\Wrapping\Utilities\TypeAccessors\InternProcessorConfig;
+use function array_key_exists;
 
 /**
  * Follows {@link PropertyPathAccessInterface} instances to check if access is
  * allowed in the context of a given root {@link TypeInterface} and maps
- * the paths according to the corresponding {@link AliasableTypeInterface::getAliases()} return.
+ * the paths according to the corresponding configured aliases.
  */
 class SchemaPathProcessor
 {
@@ -42,7 +42,7 @@ class SchemaPathProcessor
      * @param FilterableTypeInterface<TCondition, TSorting, object> $type
      * @param non-empty-list<TCondition> $conditions
      *
-     * @throws PathException Thrown if {@link AliasableTypeInterface::getAliases()} returned an invalid path.
+     * @throws PathException
      * @throws AccessException
      */
     public function mapFilterConditions(FilterableTypeInterface $type, array $conditions): void
@@ -62,7 +62,7 @@ class SchemaPathProcessor
      * @param non-empty-list<TSorting>                   $sortMethods
      *
      * @throws AccessException
-     * @throws PathException Thrown if {@link AliasableTypeInterface::getAliases()} returned an invalid path.
+     * @throws PathException
      */
     public function mapSorting(SortableTypeInterface $type, array $sortMethods): void
     {
@@ -72,75 +72,41 @@ class SchemaPathProcessor
     }
 
     /**
-     * @param non-empty-list<non-empty-string> $path
+     * Compares the given path with the {@link TransferableTypeInterface::getReadableProperties()}
+     * of the involved types.
      *
-     * @return non-empty-list<non-empty-string>
+     * @param TransferableTypeInterface<PathsBasedInterface, PathsBasedInterface, object> $type
+     * @param non-empty-list<non-empty-string> $path
      *
      * @throws PropertyAccessException
      */
-    public function mapExternReadablePath(TransferableTypeInterface $type, array $path, bool $allowAttribute): array
+    public function verifyExternReadablePath(TransferableTypeInterface $type, array $path, bool $allowAttribute): void
     {
-        $processorConfig = new ExternReadableProcessorConfig($this->typeProvider, $type, $allowAttribute);
-        $processor = $this->propertyPathProcessorFactory->createPropertyPathProcessor($processorConfig);
+        $originalPath = $path;
         try {
-            return $processor->processPropertyPath($type, [], ...$path);
+            $readableProperties = $type->getReadableProperties();
+            $readableProperties = $allowAttribute
+                ? array_merge(...$readableProperties)
+                : array_merge($readableProperties[1], $readableProperties[2]);
+
+            $pathSegment = array_shift($path);
+            if (!array_key_exists($pathSegment, $readableProperties)) {
+                $availablePropertyNames = array_keys($readableProperties);
+                throw PropertyAccessException::propertyNotAvailableInType($pathSegment, $type, $availablePropertyNames);
+            }
+
+            if ([] === $path) {
+                return;
+            }
+
+            $property = $readableProperties[$pathSegment];
+            if (!$property instanceof RelationshipReadabilityInterface) {
+                throw PropertyAccessException::nonRelationship($pathSegment, $type);
+            }
+
+            $this->verifyExternReadablePath($property->getRelationshipType(), $path, $allowAttribute);
         } catch (PropertyAccessException $exception) {
-            throw PropertyAccessException::pathDenied($type, $exception, $path);
+            throw PropertyAccessException::pathDenied($type, $exception, $originalPath);
         }
-    }
-
-    /**
-     * Check if all properties used in the sort methods are available
-     * and map the paths to be applied to the schema of the backing class.
-     *
-     * @template TCondition of PathsBasedInterface
-     * @template TSorting of PathsBasedInterface
-     *
-     * @param TypeInterface<TCondition, TSorting, object> $type
-     *
-     * @return list<TSorting>
-     *
-     * @throws PathException Thrown if {@link AliasableTypeInterface::getAliases()} returned an invalid path.
-     *
-     * @internal
-     */
-    public function processDefaultSortMethods(TypeInterface $type): array
-    {
-        $sortMethods = $type->getDefaultSortMethods();
-        if ([] === $sortMethods) {
-            return [];
-        }
-
-        $processorConfig = new InternProcessorConfig($this->typeProvider, $type);
-        $processor = $this->propertyPathProcessorFactory->createPropertyPathProcessor($processorConfig);
-        array_map([$processor, 'processPropertyPaths'], $sortMethods);
-
-        return $sortMethods;
-    }
-
-    /**
-     * Get the processed {@link TransferableTypeInterface::getAccessCondition() access condition}
-     * of the given type.
-     *
-     * @template TCondition of PathsBasedInterface
-     * @template TSorting of PathsBasedInterface
-     *
-     * @param TypeInterface<TCondition, TSorting, object> $type
-     *
-     * @return TCondition
-     *
-     * @throws AccessException
-     * @throws PathException
-     *
-     * @internal
-     */
-    public function processAccessCondition(TypeInterface $type): PathsBasedInterface
-    {
-        $condition = $type->getAccessCondition();
-        $processorConfig = new InternProcessorConfig($this->typeProvider, $type);
-        $processor = $this->propertyPathProcessorFactory->createPropertyPathProcessor($processorConfig);
-        $processor->processPropertyPaths($condition);
-
-        return $condition;
     }
 }