+++ packages/access-definitions/tests/Wrapping/Utilities/SchemaPathProcessorTest.php
@@ -16,13 +16,16 @@ use EDT\Querying\Utilities\ConditionEvaluator;
 use EDT\Querying\Utilities\Sorter;
 use EDT\Querying\Utilities\TableJoiner;
 use EDT\Wrapping\Contracts\AccessException;
+use EDT\Wrapping\Contracts\PropertyAccessException;
 use EDT\Wrapping\TypeProviders\LazyTypeProvider;
 use EDT\Wrapping\TypeProviders\PrefilledTypeProvider;
 use EDT\Wrapping\Utilities\PhpEntityVerifier;
+use EDT\Wrapping\Utilities\PropertyPathProcessor;
 use EDT\Wrapping\Utilities\PropertyPathProcessorFactory;
 use EDT\Wrapping\Utilities\SchemaPathProcessor;
-use InvalidArgumentException;
 use Tests\data\Types\AuthorType;
+use Tests\data\Types\BirthType;
+use Tests\data\Types\BookType;
 use Tests\ModelBasedTest;
 
 class SchemaPathProcessorTest extends ModelBasedTest
@@ -36,21 +39,25 @@ class SchemaPathProcessorTest extends ModelBasedTest
     protected function setUp(): void
     {
         parent::setUp();
-        $conditionFactory = new PhpConditionFactory();
         $this->sortMethodFactory = new PhpSortMethodFactory();
+        $conditionFactory = new PhpConditionFactory();
         $lazyTypeProvider = new LazyTypeProvider();
-        $propertyPathProcessorFactory = new PropertyPathProcessorFactory();
-        $schemaPathProcessor = new SchemaPathProcessor($propertyPathProcessorFactory, $lazyTypeProvider);
         $propertyAccessor = new ReflectionPropertyAccessor();
+        $propertyPathProcessorFactory = new PropertyPathProcessorFactory();
+        $this->schemaPathProcessor = new SchemaPathProcessor($propertyPathProcessorFactory, $lazyTypeProvider);
         $tableJoiner = new TableJoiner($propertyAccessor);
         $conditionEvaluator = new ConditionEvaluator($tableJoiner);
         $sorter = new Sorter($tableJoiner);
-        $entityVerifier = new PhpEntityVerifier($schemaPathProcessor, $conditionEvaluator, $sorter);
         $typeResolver = new AttributeTypeResolver();
+        $entityVerifier = new PhpEntityVerifier($this->schemaPathProcessor, $conditionEvaluator, $sorter);
+        $this->bookType = new BookType($conditionFactory, $lazyTypeProvider, $propertyAccessor, $typeResolver, $entityVerifier);
         $this->authorType = new AuthorType($conditionFactory, $lazyTypeProvider, $propertyAccessor, $entityVerifier, $typeResolver);
-        $typeProvider = new PrefilledTypeProvider([$this->authorType]);
-        $lazyTypeProvider->setAllTypes($typeProvider);
-        $this->schemaPathProcessor = new SchemaPathProcessor(new PropertyPathProcessorFactory(), $typeProvider);
+        $this->typeProvider = new PrefilledTypeProvider([
+            $this->bookType,
+            $this->authorType,
+            new BirthType($conditionFactory),
+        ]);
+        $lazyTypeProvider->setAllTypes($this->typeProvider);
     }
 
     public function testSortAccessException(): void
@@ -70,5 +77,18 @@ class SchemaPathProcessorTest extends ModelBasedTest
         self::assertCount(1, $paths);
         $path = array_pop($paths);
         self::assertSame(['name'], $path->getPath()->getAsNames());
+    }
+
+
+    public function testProcessPropertyPathWithAllowedAttribute(): void
+    {
+        $this->expectNotToPerformAssertions();
+        $this->schemaPathProcessor->verifyExternReadablePath($this->authorType, ['books', 'title'], true);
+    }
+
+    public function testProcessPropertyPathWithNonAllowedAttribute(): void
+    {
+        $this->expectException(PropertyAccessException::class);
+        $this->schemaPathProcessor->verifyExternReadablePath($this->authorType, ['books', 'title'], false);
     }
 }