+++ packages/access-definitions/tests/data/Types/AuthorType.php
@@ -10,9 +10,8 @@ use EDT\JsonApi\Properties\Relationships\PathToManyRelationshipReadability;
 use EDT\JsonApi\Properties\Relationships\PathToManyRelationshipUpdatability;
 use EDT\Querying\Contracts\PathsBasedInterface;
 use EDT\Querying\Contracts\PropertyAccessorInterface;
-use EDT\Querying\Pagination\PagePagination;
+use EDT\Querying\PropertyPaths\PropertyLink;
 use EDT\Wrapping\Contracts\TypeProviderInterface;
-use EDT\Wrapping\Contracts\Types\AliasableTypeInterface;
 use EDT\Wrapping\Contracts\Types\ExposableRelationshipTypeInterface;
 use EDT\Wrapping\Contracts\Types\FilterableTypeInterface;
 use EDT\Wrapping\Contracts\Types\IdentifiableTypeInterface;
@@ -20,7 +19,6 @@ use EDT\Wrapping\Contracts\Types\SortableTypeInterface;
 use EDT\Wrapping\Contracts\Types\TransferableTypeInterface;
 use EDT\Wrapping\Properties\IdReadabilityInterface;
 use EDT\Wrapping\Utilities\EntityVerifierInterface;
-use Pagerfanta\Pagerfanta;
 use Tests\data\Model\Person;
 
 /**
@@ -33,8 +31,7 @@ class AuthorType implements
     TransferableTypeInterface,
     FilterableTypeInterface,
     SortableTypeInterface,
-    ExposableRelationshipTypeInterface,
-    AliasableTypeInterface
+    ExposableRelationshipTypeInterface
 {
     public function __construct(
         protected readonly PathsBasedConditionFactoryInterface $conditionFactory,
@@ -70,19 +67,22 @@ class AuthorType implements
     public function getFilterableProperties(): array
     {
         return [
-            'name' => null,
-            'pseudonym' => null,
-            'books' => $this->typeProvider->requestType(BookType::class)->getInstanceOrThrow(),
-            'birthCountry' => null,
+            'name' => new PropertyLink(['name'], null),
+            'pseudonym' => new PropertyLink(['pseudonym'], null),
+            'books' => new PropertyLink(
+                ['books'],
+                $this->typeProvider->requestType(BookType::class)->getInstanceOrThrow()
+            ),
+            'birthCountry' => new PropertyLink(['birth', 'country'], null),
         ];
     }
 
     public function getSortableProperties(): array
     {
         return [
-            'name' => null,
-            'pseudonym' => null,
-            'birthCountry' => null,
+            'name' => new PropertyLink(['name'], null),
+            'pseudonym' => new PropertyLink(['pseudonym'], null),
+            'birthCountry' => new PropertyLink(['birth', 'country'], null),
         ];
     }
 
@@ -90,7 +90,6 @@ class AuthorType implements
     {
         return $this->conditionFactory->allConditionsApply(
             $this->conditionFactory->propertyHasNotSize(0, ['books']),
-            $this->conditionFactory->propertyHasNotSize(0, ['writtenBooks'])
         );
     }
 
@@ -102,14 +101,6 @@ class AuthorType implements
     public function getIdentifierFilterPath(): array
     {
         return ['name'];
-    }
-
-    public function getAliases(): array
-    {
-        return [
-            'birthCountry' => ['birth', 'country'],
-            'writtenBooks' => ['books'],
-        ];
     }
 
     public function isExposedAsRelationship(): bool
@@ -148,17 +139,6 @@ class AuthorType implements
         ];
     }
 
-    public function getInternalProperties(): array
-    {
-        return [
-            'books' => $this->typeProvider->requestType(BookType::class)->getInstanceOrThrow(),
-            'writtenBooks' => $this->typeProvider->requestType(BookType::class)->getInstanceOrThrow(),
-            'name' => null,
-            'birthCountry' => null,
-            'pseudonym' => null,
-        ];
-    }
-
     public function getIdentifier(): string
     {
         return self::class;
@@ -177,33 +157,10 @@ class AuthorType implements
                 return 0;
             }
 
-            public function isDefaultField(): bool
-            {
-                return false;
-            }
-
             public function getPropertySchema(): array
             {
                 return [];
             }
         };
-    }
-
-    public function fetchPagePaginatedEntities(
-        array $conditions,
-        array $sortMethods,
-        PagePagination $pagination
-    ): Pagerfanta {
-        throw new \Exception('Not implemented');
-    }
-
-    public function fetchEntities(array $conditions, array $sortMethods): array
-    {
-        throw new \Exception('Not implemented');
-    }
-
-    public function fetchEntity(int|string $entityIdentifier, array $conditions): ?object
-    {
-        throw new \Exception('Not implemented');
     }
 }